requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asgiref
async-timeout
attrs
backoff
bcrypt
build
cachetools
certifi
charset-normalizer
chroma-hnswlib
chromadb
click
colorama
coloredlogs
dataclasses-json
Deprecated
distro
Django
django-cors-headers
djangorestframework
djangorestframework_simplejwt
docx
drf-yasg
durationpy
exceptiongroup
faiss-cpu
fastapi
filelock
filetype
flatbuffers
frozenlist
fsspec
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-genai
google-generativeai
googleapis-common-protos
greenlet
grpcio
grpcio-status
gunicorn
h11
httpcore
httplib2
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
inflection
jiter
jsonpatch
jsonpointer
kubernetes
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-openai
langchain-text-splitters
langsmith
lxml
markdown-it-py
marshmallow
mdurl
mmh3
monotonic
mpmath
multidict
mypy-extensions
numpy
oauthlib
onnxruntime
openai
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
overrides
packaging
pillow
pip
posthog
propcache
proto-plus
protobuf
pulsar-client
pyasn1
pyasn1_modules
pydantic
pydantic_core
pydantic-settings
Pygments
PyJWT
pyparsing
pypdf
PyPika
pyproject_hooks
pyreadline3
python-dateutil
python-docx
python-dotenv
python-multipart
pytz
PyYAML
regex
requests
requests-oauthlib
requests-toolbelt
rich
rsa
setuptools
shellingham
six
sniffio
SQLAlchemy
sqlparse
starlette
sympy
tenacity
tiktoken
tokenizers
tomli
tqdm
typing_extensions
typing-inspect
tzdata
uritemplate
urllib3
uvicorn
watchfiles
websocket-client
websockets
wrapt
yarl
youtube-transcript-api
zipp
zstandard
http://127.0.0.1:8000